Move examples up
